+++ ts-fmod-plugin/consts.h
@@ -1,8 +1,0 @@
-﻿#pragma once
-
-namespace consts {
-
-    const float window_closed_volume = 0.15f;
-    const float window_open_volume = 0.3f;
-
-};