+++ ts-fmod-plugin/pch.h
@@ -21,7 +21,6 @@
 #include "fmod/fmod_studio.hpp"
 #include "fmod/fmod_errors.h"
 
-#include "consts.h"
 #include "common.h"
 
 #include "fmod_manager.h"